Changed auto_now to auto_now_add.
This changed was done because auto_now is updated every time the models instance is updated, and that is not what we wanted it to do. The correct behavior is that these datetime fields are only updated when they are created, and that is what auto_now_add allows us to do.